Renamed to PWVSpring & added extra functioniality
The component was renamed from PlayWhenVisible to PWVSpring, where PWV
is an alias for PlayWhenVisible.

Extra functionaltiy includes the onRest and onStart methods that are
triggered when the animation starts and stops.

Additionally, the component now uses a child function that allows the
user to pass the animation to the style prop in any animated component.